+++ routes/api/contacts.js
@@ -1,24 +1,12 @@
 const express = require('express');
 
-const Joi = require("joi");
-
 const contacts = require('../../models/contacts');
 
 const { HttpError } = require('../../helpers');
 
-const router = express.Router()
+const {addSchema} = require('../../schemas/contacts')
 
-const addSchema = Joi.object({
-  name: Joi.string().required().messages({
-		'any.required': `missing name field`
-    }),
-  email: Joi.string().required().messages({
-		'any.required': `missing email field`
-    }),
-  phone: Joi.required().messages({
-		'any.required': `missing phone field`
-    }),
-})
+const router = express.Router()
 
 router.get('/', async (req, res, next) => {
   try {
